+++ tools/amavlink/release_notes.md
@@ -4,6 +4,7 @@
 
 * Parameters:
     * Added get_number_of_params.
+    * Get parameters by index.
 * Capture strmatch messages to stdout.
 
 ## v 0.06